+++ src/shared/ui/components/index.ts
@@ -1,3 +1,2 @@
 export {default as CategoryBar} from './CategoryBar';
-export {default as ReviewCard} from './ReviewCard';
 export {default as Toast, type ToastProps} from './Toast';
